Servant.Foreign

Description

Generalizes all the data needed to make code generation work with arbitrary programming languages.

See documentation of HasForeignType for a simple example. listFromAPI returns a list of all your endpoints and their foreign types, given a mapping from Haskell types to foreign types (conventionally called ftypes below).

Main API

listFromAPI :: (HasForeign lang ftype api, GenerateList ftype (Foreign ftype api)) => Proxy lang -> Proxy ftype -> Proxy api -> [Req ftype]

Generate the necessary data for codegen as a list, each Req describing one endpoint from your API type.

data Req ftype

Full description of an endpoint in your API, generated by listFromAPI. It should give you all the information needed to generate foreign language bindings.

Every field containing ftype will use the foreign type mapping specified via HasForeignType (see its docstring on how to set that up).

See https://docs.servant.dev/en/stable/tutorial/ApiType.html for accessible documentation of the possible content of an endpoint.

Constructors

Req

Fields

_reqUrl :: Url ftype

Full list of URL segments, including captures

_reqMethod :: Method

"GET"/"POST"/"PUT"/…

_reqHeaders :: [HeaderArg ftype]

Headers required by this endpoint, with their type

_reqBody :: Maybe ftype

Foreign type of the expected request body (ReqBody), if any

_reqReturnType :: Maybe ftype

The foreign type of the response, if any

_reqFuncName :: FunctionName

The URL segments rendered in a way that they can be easily concatenated into a canonical function name

_reqBodyContentType :: ReqBodyContentType

The content type the request body is transferred as.

This is a severe limitation of servant-foreign currently, as we only allow the content type to be JSON no user-defined content types. (ReqBodyMultipart is not actually implemented.)

Thus, any routes looking like this will work:

"foo" :> Get '[JSON] Foo

while routes like

"foo" :> Get '[MyFancyContentType] Foo

will fail with an error like

• JSON expected in list '[MyFancyContentType]

defReq :: Req ftype

class HasForeignType lang ftype a where

HasForeignType maps Haskell types with types in the target language of your backend. For example, let's say you're implementing a backend to some language X, and you want a Text representation of each input/output type mentioned in the API:

-- First you need to create a dummy type to parametrize your
-- instances.
data LangX

-- Otherwise you define instances for the types you need
instance HasForeignType LangX Text Int where
   typeFor _ _ _ = "intX"

-- Or for example in case of lists
instance HasForeignType LangX Text a => HasForeignType LangX Text [a] where
   typeFor lang type _ = "listX of " <> typeFor lang ftype (Proxy :: Proxy a)

Finally to generate list of information about all the endpoints for an API you create a function of a form:

getEndpoints :: (HasForeign LangX Text api, GenerateList Text (Foreign Text api))
             => Proxy api -> [Req Text]
getEndpoints api = listFromAPI (Proxy :: Proxy LangX) (Proxy :: Proxy Text) api

-- If language __X__ is dynamically typed then you can use
-- a predefined NoTypes parameter with the NoContent output type:

getEndpoints :: (HasForeign NoTypes NoContent api, GenerateList Text (Foreign NoContent api))
             => Proxy api -> [Req NoContent]
getEndpoints api = listFromAPI (Proxy :: Proxy NoTypes) (Proxy :: Proxy NoContent) api

Methods

typeFor :: Proxy lang -> Proxy ftype -> Proxy a -> ftype

class GenerateList ftype reqs where

Utility class used by listFromAPI which computes the data needed to generate a function for each endpoint and hands it all back in a list.

Methods

generateList :: reqs -> [Req ftype]

class HasForeign lang ftype (api :: *) where

Implementation of the Servant framework types.

Relevant instances: Everything containing HasForeignType.

Associated Types

type Foreign ftype api :: *

Methods

foreignFor :: Proxy lang -> Proxy ftype -> Proxy api -> Req ftype -> Foreign ftype api

data NoTypes

The language definition without any foreign types. It can be used for dynamic languages which do not do type annotations.

Subtypes of Req

data Url ftype

Full endpoint url, with all captures and parameters

Constructors

Url

Fields

_path :: Path ftype

Url path, list of either static segments or captures

"foo/{id}/bar"

_queryStr :: [QueryArg ftype]

List of query args

"?foo=bar&a=b"

_frag :: Maybe ftype

Url fragment.

Not sent to the HTTP server, so only useful for frontend matters (e.g. inter-page linking).

#fragmentText

type Path ftype = [Segment ftype]

newtype Segment ftype

A part of the Url’s path.

Constructors

Segment

Fields

unSegment :: SegmentType ftype

data SegmentType ftype

Constructors

Static PathSegment
Static path segment.

"foo/bar/baz"

contains the static segments "foo", "bar" and "baz".

Cap (Arg ftype)
A capture.

"user/{userid}/name"

would capture the arg userid with type ftype.

isCapture :: Segment ftype -> Bool

Whether a segment is a Cap.

captureArg :: Segment ftype -> Arg ftype

Crashing Arg extraction from segment, TODO: remove

data QueryArg ftype

Url Query argument.

Urls can contain query arguments, which is a list of key-value pairs. In a typical url, query arguments look like this:

?foo=bar&alist[]=el1&alist[]=el2&aflag

Each pair can be

?foo=bar: a plain key-val pair, either optional or required (QueryParam)

?aflag: a flag (no value, implicitly Bool with default false if it’s missing) (QueryFlag)

?alist[]=el1&alist[]=el2: list of values (QueryParams)

_queryArgType will be set accordingly.

For the plain key-val pairs (QueryParam), _queryArgName’s ftype will be wrapped in a Maybe if the argument is optional.

Constructors

QueryArg

Fields

_queryArgName :: Arg ftype

Name and foreign type of the argument. Will be wrapped in Maybe if the query is optional and in a `[]` if the query is a list

_queryArgType :: ArgType

one of normal/plain, list or flag

data ArgType

Type of a QueryArg.

Constructors

Normal

Flag

List

data HeaderArg ftype

Constructors

HeaderArg
The name of the header and the foreign type of its value.

Fields

_headerArg :: Arg ftype

ReplaceHeaderArg
Unused, will never be set.

TODO: remove

Fields

_headerArg :: Arg ftype

_headerPattern :: Text

data Arg ftype

Maps a name to the foreign type that belongs to the annotated value.

Used for header args, query args, and capture args.

Constructors

Arg

Fields

_argName :: PathSegment

The name to be captured.

Only for capture args it really denotes a path segment.

_argType :: ftype

Foreign type the associated value will have

newtype FunctionName

Canonical name of the endpoint, can be used to generate a function name.

You can use the functions in Servant.Foreign.Inflections, like camelCase to transform to Text.

Constructors

FunctionName

Fields

unFunctionName :: [Text]

data ReqBodyContentType

See documentation of _reqBodyContentType

Constructors

ReqBodyJSON

ReqBodyMultipart

newtype PathSegment

See documentation of Arg

Constructors

PathSegment

Fields

unPathSegment :: Text

Lenses

argName :: forall ftype. Lens' (Arg ftype) PathSegment

argType :: forall ftype ftype. Lens (Arg ftype) (Arg ftype) ftype ftype

argPath :: Getter (Arg ftype) Text

reqUrl :: forall ftype. Lens' (Req ftype) (Url ftype)

reqMethod :: forall ftype. Lens' (Req ftype) Method

reqHeaders :: forall ftype. Lens' (Req ftype) [HeaderArg ftype]

reqBody :: forall ftype. Lens' (Req ftype) (Maybe ftype)

reqBodyContentType :: forall ftype. Lens' (Req ftype) ReqBodyContentType

reqReturnType :: forall ftype. Lens' (Req ftype) (Maybe ftype)

reqFuncName :: forall ftype. Lens' (Req ftype) FunctionName

path :: forall ftype. Lens' (Url ftype) (Path ftype)

queryStr :: forall ftype. Lens' (Url ftype) [QueryArg ftype]

queryArgName :: forall ftype ftype. Lens (QueryArg ftype) (QueryArg ftype) (Arg ftype) (Arg ftype)

queryArgType :: forall ftype. Lens' (QueryArg ftype) ArgType

headerArg :: forall ftype ftype. Lens (HeaderArg ftype) (HeaderArg ftype) (Arg ftype) (Arg ftype)

Prisms

_PathSegment :: Iso' PathSegment Text

_HeaderArg :: forall ftype. Prism' (HeaderArg ftype) (Arg ftype)

_ReplaceHeaderArg :: forall ftype. Prism' (HeaderArg ftype) (Arg ftype, Text)

_Static :: forall ftype. Prism' (SegmentType ftype) PathSegment

_Cap :: forall ftype ftype. Prism (SegmentType ftype) (SegmentType ftype) (Arg ftype) (Arg ftype)

_Normal :: Prism' ArgType ()

_Flag :: Prism' ArgType ()

_List :: Prism' ArgType ()

Re-exports

module Servant.API

module Servant.Foreign.Inflections