Servant.Foreign.Inflections

Documentation

concatCase :: FunctionName -> Text

Simply concat each part of the FunctionName together.

[ "get", "documents", "by", "id" ] → "getdocumentsbyid"

snakeCase :: FunctionName -> Text

Use the snake_case convention. Each part is separated by a single underscore character.

[ "get", "documents", "by", "id" ] → "get_documents_by_id"

camelCase :: FunctionName -> Text

Use the camelCase convention. The first part is lower case, every other part starts with an upper case character.

[ "get", "documents", "by", "id" ] → "getDocumentsById"

concatCaseL :: Getter FunctionName Text

snakeCaseL :: Getter FunctionName Text

camelCaseL :: Getter FunctionName Text