Servant.Foreign

Description

Generalizes all the data needed to make code generation work with arbitrary programming languages.

Documentation

class HasForeign layout where

Associated Types

type Foreign layout :: *

Methods

foreignFor :: Proxy layout -> Req -> Foreign layout

newtype Segment

Constructors

Segment

Fields

_segment :: SegmentType

data SegmentType

Constructors

Static Text
a static path segment. like "/foo"

Cap Arg
a capture. like "/:userid"

type FunctionName = [Text]

data QueryArg

Constructors

QueryArg

Fields

_argName :: Arg

_argType :: ArgType

data HeaderArg

Constructors

HeaderArg

Fields

headerArgName :: Text

ReplaceHeaderArg

Fields

headerArgName :: Text

headerPattern :: Text

data ArgType

Constructors

Normal

Flag

List

data Req

captureArg :: Segment -> Arg

defReq :: Req

concatCase :: FunctionName -> Text

Function name builder that simply concat each part together

snakeCase :: FunctionName -> Text

Function name builder using the snake_case convention. each part is separated by a single underscore character.

camelCase :: FunctionName -> Text

Function name builder using the CamelCase convention. each part begins with an upper case character.

argType :: Lens' QueryArg ArgType

argName :: Lens' QueryArg Arg

isCapture :: Segment -> Bool

funcName :: Lens' Req FunctionName

path :: Lens' Url Path

reqUrl :: Lens' Req Url

reqBody :: Lens' Req Bool

reqHeaders :: Lens' Req [HeaderArg]

reqMethod :: Lens' Req Method

segment :: Iso' Segment SegmentType

queryStr :: Lens' Url [QueryArg]

module Servant.API